- Adiciona método RegistrarAlbum com validação de banda existente. - Adiciona método RegistrarMusica com validação de gênero e vínculo opcional com álbum. - Implementa método utilitário ConverterTempoParaSegundos para aceitar formatos "mm:ss" e minutos inteiros. - Cria método RetornarAlbumPorNome na classe Banda com busca case-insensitive (ToUpper). - Corrige bugs de formatação na leitura da duração da música.

+++ PrimeiroProjeto/Musica.cs
@@ -5,14 +5,14 @@ public class Musica
     // Propriedades da classe Musica
     public string NomeDaMusica { get; }
     public Banda Artista { get; }
-    public double Duracao { get; set; }
+    public int Duracao { get; set; }
     public bool Disponivel { get; set; }
     public Genero TipoDeGenero { get; set; }
     public string DescricaoResumida =>
         $"A música '{NomeDaMusica}' pertence ao artista {Artista.NomeDaBanda}";
 
     // Cunstrutor
-    public Musica(Banda banda, string nomeDaMusica, double duracao, bool disponivel, Genero genero)
+    public Musica(Banda banda, string nomeDaMusica, int duracao, bool disponivel, Genero genero)
     {
         Artista = banda;
         NomeDaMusica = nomeDaMusica;
